bigint c++98

+++ practice/lvl1/bigint/main.cpp
@@ -1,6 +1,9 @@
 #include "bigint.hpp"
 #include <iostream>
 
+#include "bigint.hpp"
+#include <iostream>
+
 int main()
 {
     const bigint a(42);
@@ -19,7 +22,7 @@ int main()
     std::cout << "b++ = " << b++ << std::endl;
     std::cout << "(b << 10) + 42 = " <<  (b << 10) + 42 << std::endl;
     std::cout << "(d <<= 4) = " << (d <<= 4) << std::endl;
-    std::cout << "(d >>= 2) = " << (d >>= 2) << std::endl;
+    std::cout << "(d >>= 2) = " << (d >>= (const bigint)2) << std::endl;
     std::cout << "a = " << a << std::endl;
     std::cout << "d = " << d << std::endl;
     bigint f("42");
